+++ scripts/run_demo.sh
@@ -41,19 +41,16 @@ curl -s -i -X 'POST' \
   "name": "Tenant1"
 }'
 
-echo "Setting up policy that allows reading and creating entities under tenant Tenant1 and path / ..."
+echo "Setting up tenant Tenant2..."
 curl -s -i -X 'POST' \
-'http://127.0.0.1:8085/v1/policies/' \
--H 'accept: */*' \
--H 'fiware_service: Tenant1' \
--H 'fiware_service_path: /' \
--H 'Content-Type: application/json' \
--d '{
-"access_to": "*",
-"resource_type": "entity",
-"mode": ["acl:Read"],
-"agent": ["acl:AuthenticatedAgent"]
+  'http://127.0.0.1:8085/v1/tenants/' \
+  -H 'accept: */*' \
+  -H 'Content-Type: application/json' \
+  -d '{
+  "name": "Tenant2"
 }'
+
+echo "Setting up policy that allows creating entities under tenant Tenant1 and path / ..."
 
 curl -s -i -X 'POST' \
 'http://127.0.0.1:8085/v1/policies/' \